requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pillow>=10.1.0
transformers==4.36.2
torch==2.6.0
python-multipart==0.0.6
pydantic==2.5.0
openai>=1.0.0 